backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Celery (Background Tasks)
celery==5.3.4
flower==2.0.1

# MongoDB (Optional)
motor==3.3.2
pymongo==4.6.1

# ML & Data Science
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2

# TensorFlow (Optional - for deep learning)
# tensorflow==2.15.0

# XGBoost
xgboost==2.0.3

# HTTP Requests
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring
prometheus-client==0.19.0
